+++ dev/start.sh
@@ -41,6 +41,7 @@ export CTAGS_PROCESSES=1
 export SRC_SYNTECT_SERVER=http://localhost:9238
 export SRC_FRONTEND_INTERNAL=localhost:3090
 export SRC_PROF_HTTP=
+export SRC_PROF_SERVICES=$(cat dev/src-prof-services.json)
 export NPM_CONFIG_LOGLEVEL=silent
 
 # To use webpack-dev-server for auto-reloading, use: