+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Bug fixes
 
 * More detailed logging to help diagnose errors with third-party authentication providers.
 * Anchors (such as `#my-section`) in rendered Markdown files are now supported.
+* Instrumentation section for admins. For each service we expose pprof, prometheus metrics and traces.
 
 ### Bug fixes